Refactored and removed extra files
