+++ fabric-dehardcoding-api-v0/src/main/java/net/fabricmc/fabric/api/dehardcoding/item/v1/FabricCrossbowHooks.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.item.v1;
+package net.fabricmc.fabric.api.dehardcoding.item.v1;
 
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.item.ArrowItem;
