+++ fabric-dehardcoding-api-v0/src/main/java/net/fabricmc/fabric/mixin/dehardcoding/AbstractSkeletonEntityMixin.java
@@ -14,27 +14,22 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.tool.client;
+package net.fabricmc.fabric.mixin.dehardcoding;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.client.render.entity.model.SkeletonEntityModel;
+import net.minecraft.entity.mob.AbstractSkeletonEntity;
 import net.minecraft.item.BowItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 
-@Mixin(SkeletonEntityModel.class)
-public class SkeletonEntityModelMixin {
-	@Redirect(method = "animateModel", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;getItem()Lnet/minecraft/item/Item;"))
-	private Item animateModel(ItemStack itemStack) {
-		return itemStack.getItem() instanceof BowItem ? Items.BOW : itemStack.getItem();
-	}
-
-	@Redirect(method = "setAngles", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;getItem()Lnet/minecraft/item/Item;"))
-	private Item setAngles(ItemStack itemStack) {
+@Mixin(AbstractSkeletonEntity.class)
+public class AbstractSkeletonEntityMixin {
+	@Redirect(method = "updateAttackType", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;getItem()Lnet/minecraft/item/Item;"))
+	private Item updateAttackType(ItemStack itemStack) {
 		return itemStack.getItem() instanceof BowItem ? Items.BOW : itemStack.getItem();
 	}
 }
