+++ fabric-dehardcoding-api-v0/src/main/java/net/fabricmc/fabric/mixin/dehardcoding/EntityBowUseMixin.java
@@ -14,22 +14,23 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.tool;
+package net.fabricmc.fabric.mixin.dehardcoding;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import net.minecraft.entity.mob.AbstractSkeletonEntity;
 import net.minecraft.item.BowItem;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
+import net.minecraft.item.RangedWeaponItem;
 
-@Mixin(AbstractSkeletonEntity.class)
-public class AbstractSkeletonEntityMixin {
-	@Redirect(method = "updateAttackType", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;getItem()Lnet/minecraft/item/Item;"))
-	private Item updateAttackType(ItemStack itemStack) {
-		return itemStack.getItem() instanceof BowItem ? Items.BOW : itemStack.getItem();
+@Mixin({ AbstractSkeletonEntity.class })
+public class EntityBowUseMixin {
+	@Inject(method = "canUseRangedWeapon", at = @At("HEAD"))
+	public void canUseRangedWeapon(RangedWeaponItem weapon, CallbackInfoReturnable<Boolean> cir) {
+		if (weapon instanceof BowItem) {
+			cir.setReturnValue(true);
+		}
 	}
 }
