+++ fabric-dehardcoding-api-v0/src/main/java/net/fabricmc/fabric/mixin/dehardcoding/EntityCrossbowUseMixin.java
@@ -14,22 +14,23 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.tool;
+package net.fabricmc.fabric.mixin.dehardcoding;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import net.minecraft.entity.mob.AbstractSkeletonEntity;
-import net.minecraft.item.BowItem;
+import net.minecraft.entity.mob.PiglinEntity;
+import net.minecraft.entity.mob.PillagerEntity;
+import net.minecraft.item.CrossbowItem;
 import net.minecraft.item.RangedWeaponItem;
 
-@Mixin({ AbstractSkeletonEntity.class })
-public class EntityBowUseMixin {
+@Mixin({ PiglinEntity.class, PillagerEntity.class })
+public class EntityCrossbowUseMixin {
 	@Inject(method = "canUseRangedWeapon", at = @At("HEAD"))
 	public void canUseRangedWeapon(RangedWeaponItem weapon, CallbackInfoReturnable<Boolean> cir) {
-		if (weapon instanceof BowItem) {
+		if (weapon instanceof CrossbowItem) {
 			cir.setReturnValue(true);
 		}
 	}
