+++ fabric-dehardcoding-api-v0/src/main/java/net/fabricmc/fabric/mixin/dehardcoding/ProjectileUtilMixin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.tool;
+package net.fabricmc.fabric.mixin.dehardcoding;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
