+++ fabric-dehardcoding-api-v0/src/main/java/net/fabricmc/fabric/mixin/dehardcoding/client/HeldItemRendererMixin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.tool.client;
+package net.fabricmc.fabric.mixin.dehardcoding.client;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
